Add user-defined functions. Update docstrings
To pretty-print user-defined functions, it's necessary to extract the
sub-expressions first, then create strings for each, and only then
escape the names of the functions. If the functions were escaped before
the strings are created, the strings would be hard to read, e.g.
`:($(Expr(:escape, :foo)))` instead of `foo`.